% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_1.R
\name{add_1}
\alias{add_1}
\title{Add 1}
\usage{
add_1(x)
}
\arguments{
\item{x}{is <numeric> (int or dbl)}
}
\value{
a number that is x + 1 <numeric>
}
\description{
Add 1 to any number
}
\examples{
add_1(3)
# [1] 4
}
