% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complicated.R
\name{complicated}
\alias{complicated}
\title{Complicated}
\usage{
complicated(df, func = "mean")
}
\arguments{
\item{df}{is a tibble}

\item{x}{is a numeric column within df}
}
\value{
a tibble with calculated field
}
\description{
Performs a calculation based off of input
}
\examples{
my_tib = tibble::tibble(x = c(1,10,39))
complicated(a)
# A tibble: 3 x 2
x     mean_x
<dbl>  <dbl>
 1    16.7
 10   16.7
 39   16.7
}
