\name{scott-package}
\alias{scott-package}
\alias{scott}
\docType{package}
\title{
\packageTitle{scott}
}
\description{
\packageDescription{scott}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scott}
\packageIndices{scott}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{scott}

Maintainer: \packageMaintainer{scott}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
