% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scott0.R
\name{scott}
\alias{scott}
\title{Simulate a coalescent transmission tree}
\usage{
scott(tfgy, s, X, min_dt = NULL, max_dt = NULL, event_tol = 1)
}
\arguments{
\item{tfgy}{output of demo/epi simulation}

\item{s}{vector of sample times}

\item{X}{matrix of sample states (probabilities)}
}
\value{
A phylogenetic tree which contains a $waifw attribute. This is a data frame containing transmission pairs and time of transmission.
}
\description{
This function will return a phylogeny in the `ape` package format with additional attributes showing who infected whom.
}
